Modularize Different Components into directories
