+++ libhdhomerun/channel.h
