+++ qt/hdhomerun-gui_en_US.ts
