+++ qt/mainwindow.h
